Type BatchError model
Done
